Move getAllNested and getNested from FolderService to VaultFilterService
