+++ libs/common/src/abstractions/folder/folder.service.abstraction.ts
@@ -10,8 +10,6 @@ export abstract class FolderService {
   get: (id: string) => Promise<Folder>;
   getAll: () => Promise<Folder[]>;
   getAllDecrypted: () => Promise<FolderView[]>;
-  getAllNested: (folders?: FolderView[]) => Promise<TreeNode<FolderView>[]>;
-  getNested: (id: string) => Promise<TreeNode<FolderView>>;
 }
 
 export abstract class InternalFolderService extends FolderService {
